requirements.txt
djangorestframework==3.14.0
django
django-rest-auth==0.9.5
django-allauth==0.54.0
psycopg2
